+++ docs/CHANGELOG.md
@@ -1,3 +1,30 @@
+<a name="3.1.8"></a>
+# 3.1.8 (2019-10-18)
+
+### Features
+* **core:** added default font to presets.
+* **editor:** keep label value when using text type field.
+
+### Bug fixes
+* **editor:** fixed some of the fonts not being rendered correctly on the pdf
+* **tools:** schema service now recognises lists with enum
+* **tools:** fix schema service only picking first object from an array of objects
+
+<a name="3.1.7"></a>
+# 3.1.7 (2019-10-09)
+
+* **tools:** fix schema service only picking first object from an array of objects
+
+<a name="3.1.6"></a>
+# 3.1.6 (2019-10-08)
+
+### Features
+* **editor:** when opening a setup will now try to update the label to match current language.
+
+### Bug fixes
+* **editor:** when loading setup merge the defaultAvailableFields mappings to the loaded data 
+* **editor:** fixed few typos 
+
 <a name="3.1.5"></a>
 # 3.1.5 (2019-09-05)